project/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
bcrypt==4.0.1
billiard==4.1.0
bleach==6.0.0
blinker==1.6.2
Brotli==1.0.9
cachelib==0.9.0
celery==5.3.1
cffi==1.15.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.1.0
cssselect2==0.7.0
cycler==0.11.0
dnspython==2.4.1
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.0
Flask-SQLAlchemy==3.0.5
Flask-SSE==1.0.0
Flask-WTF==1.1.1
fonttools==4.41.1
gevent==23.7.0
greenlet==2.0.2
gunicorn==21.2.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.3.1
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.2
packaging==23.1
passlib==1.7.4
Pillow==10.0.0
prompt-toolkit==3.0.39
pycparser==2.21
pydyf==0.7.0
pyparsing==3.0.9
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
six==1.16.0
SQLAlchemy==2.0.19
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
Werkzeug==2.3.6
WTForms==3.0.1
zope.event==5.0
zope.interface==6.0
zopfli==0.2.2
